refactor(core): 💡 clarifying the logic of how a parent selects its children

+++ widgets/src/label.rs
@@ -1,10 +1,11 @@
 use ribir_core::prelude::*;
 
+#[derive(ChildOfCompose)]
 pub struct Label(pub DeclareInit<CowArc<str>>);
 
 impl Label {
   #[inline]
-  pub fn new<const M: u8>(str: impl DeclareInto<CowArc<str>, M>) -> Self {
+  pub fn new<const M: usize>(str: impl DeclareInto<CowArc<str>, M>) -> Self {
     Self(str.declare_into())
   }
 }
